feat(frontend): reorganización y optimización de la estructura
- Centralización de metadata en carpeta dedicada
- Organización de componentes y estructura de archivos
- Mejoras en el diseño y la interfaz de usuario
- Documentación con README para la carpeta de metadata
- Optimización de importaciones y exportaciones
