+++ pixela-frontend/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "next": "14.2.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.5.0"
   },
   "devDependencies": {
     "@types/node": "^20.17.30",
