+++ pixela-frontend/src/app/components/series/SeriesGrid.tsx
@@ -1,0 +1,35 @@
+'use client';
+
+import { Series } from "@/lib/interface/series/trending-series";
+import Image from "next/image";
+
+interface Props {
+    series: Series[];
+}
+
+// TODO Mejorar responsive para que se pueda ver en cualquier dispositivo
+export const SeriesGrid = ({ series }: Props) => { 
+    return (
+        <div className="flex justify-center items-center min-h-screen bg-[#1E1E1E]">
+            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-6 gap-3 p-4">
+                {series.map((serie) => (
+                    <div 
+                        key={serie.id} 
+                        className="relative overflow-hidden rounded-lg shadow-2xl mx-auto"
+                    >
+                        {/* Imagen del Poster */}
+                        <div className="relative h-full w-full">
+                            <Image 
+                                src={`https://image.tmdb.org/t/p/w500${serie.poster_path}`} 
+                                alt={serie.title} 
+                                className="object-cover"
+                                width={297}
+                                height={437}
+                            />
+                        </div>
+                    </div>
+                ))}
+            </div>
+        </div>
+    );
+};
