+++ pixela-frontend/src/app/series/index.ts
@@ -1,0 +1,3 @@
+
+//TODO Exportar todos los componentes de la carpeta series
+export * from '../components/series/SeriesGrid';
