+++ pixela-frontend/src/app/series/layout.tsx
