+++ pixela-frontend/src/lib/api/index.ts
@@ -1,0 +1,2 @@
+//TODOExportar todas las funciones de API desde un punto central
+export * from './series/series';
