+++ pixela-frontend/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './api';
+export * from './interface/series/trending-series';
