+++ pixela-frontend/src/lib/interface/series/trending-series.ts
