+++ pixela-frontend/src/metadata/index.ts
@@ -1,0 +1,1 @@
+export * from './pages/series';
