+++ pixela-frontend/src/metadata/pages/series.ts
@@ -1,0 +1,4 @@
+export const seriesMetadata = {
+    title: 'Series',
+    description: 'Series de televisión',
+}
